+++ database/gov.go
@@ -3,84 +3,36 @@ package database
 import (
 	"encoding/json"
 	"fmt"
+	"strings"
 	"time"
 
+	"github.com/cosmos/cosmos-sdk/codec"
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
 	govtypesv1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
-	"github.com/gogo/protobuf/proto"
-
-	govtypesv1beta1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1beta1"
 	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
+	"github.com/lib/pq"
+
 	dbtypes "github.com/forbole/bdjuno/v4/database/types"
 	"github.com/forbole/bdjuno/v4/types"
-	"github.com/lib/pq"
 )
 
 // SaveGovParams saves the given x/gov parameters inside the database
 func (db *Db) SaveGovParams(params *types.GovParams) error {
-
-	depositParamsBz, err := json.Marshal(&params.DepositParams)
+	paramsBz, err := json.Marshal(&params.Params)
 	if err != nil {
-		return fmt.Errorf("error while marshaling deposit params: %s", err)
-	}
-
-	votingParamsBz, err := json.Marshal(&params.VotingParams)
-	if err != nil {
-		return fmt.Errorf("error while marshaling voting params: %s", err)
-	}
-
-	tallyingParams, err := json.Marshal(&params.TallyParams)
-	if err != nil {
-		return fmt.Errorf("error while marshaling tally params: %s", err)
+		return fmt.Errorf("error while marshalling gov params: %s", err)
 	}
 
 	stmt := `
-INSERT INTO gov_params(deposit_params, voting_params, tally_params, height)
-VALUES ($1, $2, $3, $4)
+INSERT INTO gov_params(params, height) 
+VALUES ($1, $2) 
 ON CONFLICT (one_row_id) DO UPDATE
-	SET deposit_params = excluded.deposit_params,
-  		voting_params = excluded.voting_params,
-		tally_params = excluded.tally_params,
+	SET params = excluded.params,
 		height = excluded.height
 WHERE gov_params.height <= excluded.height`
-	_, err = db.SQL.Exec(stmt, string(depositParamsBz), string(votingParamsBz), string(tallyingParams), params.Height)
+	_, err = db.SQL.Exec(stmt, string(paramsBz), params.Height)
 	if err != nil {
 		return fmt.Errorf("error while storing gov params: %s", err)
-	}
-
-	return nil
-}
-
-// SaveGenesisGovParams saves the genesis x/gov parameters inside the database
-func (db *Db) SaveGenesisGovParams(params *types.GenesisGovParams) error {
-
-	depositParamsBz, err := json.Marshal(&params.DepositParams)
-	if err != nil {
-		return fmt.Errorf("error while marshaling genesis deposit params: %s", err)
-	}
-
-	votingParamsBz, err := json.Marshal(&params.VotingParams)
-	if err != nil {
-		return fmt.Errorf("error while marshaling genesis voting params: %s", err)
-	}
-
-	tallyingParams, err := json.Marshal(&params.TallyParams)
-	if err != nil {
-		return fmt.Errorf("error while marshaling genesis tally params: %s", err)
-	}
-
-	stmt := `
-INSERT INTO gov_params(deposit_params, voting_params, tally_params, height)
-VALUES ($1, $2, $3, $4)
-ON CONFLICT (one_row_id) DO UPDATE
-	SET deposit_params = excluded.deposit_params,
-  		voting_params = excluded.voting_params,
-		tally_params = excluded.tally_params,
-		height = excluded.height
-WHERE gov_params.height <= excluded.height`
-	_, err = db.SQL.Exec(stmt, string(depositParamsBz), string(votingParamsBz), string(tallyingParams), params.Height)
-	if err != nil {
-		return fmt.Errorf("error while storing genesis gov params: %s", err)
 	}
 
 	return nil
@@ -100,28 +52,13 @@ func (db *Db) GetGovParams() (*types.GovParams, error) {
 
 	row := rows[0]
 
-	var depositParams types.DepositParams
-	err = json.Unmarshal([]byte(row.DepositParams), &depositParams)
+	var params govtypesv1.Params
+	err = json.Unmarshal([]byte(row.Params), &params)
 	if err != nil {
 		return nil, err
 	}
 
-	var votingParams types.VotingParams
-	err = json.Unmarshal([]byte(row.VotingParams), &votingParams)
-	if err != nil {
-		return nil, err
-	}
-
-	var tallyParams types.TallyParams
-	err = json.Unmarshal([]byte(row.TallyParams), &tallyParams)
-	if err != nil {
-		return nil, err
-	}
-
-	return types.NewGovParams(
-		votingParams, depositParams, tallyParams,
-		row.Height,
-	), nil
+	return types.NewGovParams(&params, row.Height), nil
 }
 
 // --------------------------------------------------------------------------------------------------------------------
@@ -136,7 +73,7 @@ func (db *Db) SaveProposals(proposals []types.Proposal) error {
 
 	proposalsQuery := `
 INSERT INTO proposal(
-	id, title, description, content, proposer_address, proposal_route, proposal_type, status,
+	id, title, description, metadata, content, proposer_address, status,
     submit_time, deposit_end_time, voting_start_time, voting_end_time
 ) VALUES`
 	var proposalsParams []interface{}
@@ -146,34 +83,27 @@ INSERT INTO proposal(
 		accounts = append(accounts, types.NewAccount(proposal.Proposer))
 
 		// Prepare the proposal query
-		vi := i * 12
-		proposalsQuery += fmt.Sprintf("($%d,$%d,$%d,$%d,$%d,$%d,$%d,$%d,$%d,$%d,$%d,$%d),",
-			vi+1, vi+2, vi+3, vi+4, vi+5, vi+6, vi+7, vi+8, vi+9, vi+10, vi+11, vi+12)
+		vi := i * 11
+		proposalsQuery += fmt.Sprintf("($%d,$%d,$%d,$%d,$%d,$%d,$%d,$%d,$%d,$%d,$%d),",
+			vi+1, vi+2, vi+3, vi+4, vi+5, vi+6, vi+7, vi+8, vi+9, vi+10, vi+11)
 
-		// Encode the content properly
-		protoContent, ok := proposal.Content.(proto.Message)
-		if !ok {
-			return fmt.Errorf("invalid proposal content types: %T", proposal.Content)
-		}
-
-		anyContent, err := codectypes.NewAnyWithValue(protoContent)
-		if err != nil {
-			return fmt.Errorf("error while wrapping proposal proto content: %s", err)
-		}
-
-		contentBz, err := db.EncodingConfig.Codec.MarshalJSON(anyContent)
-		if err != nil {
-			return fmt.Errorf("error while marshaling proposal content: %s", err)
+		var jsonMessages []string
+		var protoCodec codec.ProtoCodec
+		for _, msg := range proposal.Messages {
+			contentBz, err := protoCodec.MarshalJSON(msg)
+			if err != nil {
+				return fmt.Errorf("error while marshalling proposal msg: %s", err)
+			}
+			jsonMessages = append(jsonMessages, string(contentBz))
 		}
 
 		proposalsParams = append(proposalsParams,
-			proposal.ProposalID,
-			proposal.Content.GetTitle(),
-			proposal.Content.GetDescription(),
-			string(contentBz),
+			proposal.ID,
+			proposal.Title,
+			proposal.Summary,
+			proposal.Metadata,
+			fmt.Sprintf("[%s]", strings.Join(jsonMessages, ",")),
 			proposal.Proposer,
-			proposal.ProposalRoute,
-			proposal.ProposalType,
 			proposal.Status,
 			proposal.SubmitTime,
 			proposal.DepositEndTime,
@@ -182,7 +112,7 @@ INSERT INTO proposal(
 		)
 	}
 
-	// Store the accounts
+	// Store the proposers accounts
 	err := db.SaveAccounts(accounts)
 	if err != nil {
 		return fmt.Errorf("error while storing proposers accounts: %s", err)
@@ -190,7 +120,17 @@ INSERT INTO proposal(
 
 	// Store the proposals
 	proposalsQuery = proposalsQuery[:len(proposalsQuery)-1] // Remove trailing ","
-	proposalsQuery += " ON CONFLICT DO NOTHING"
+	proposalsQuery += `
+ON CONFLICT (id) DO UPDATE 
+	SET title = excluded.title,
+  		description = excluded.description,
+		content = excluded.content,
+		proposer_address = excluded.proposer_address,
+		status = excluded.status,
+		submit_time = excluded.submit_time,
+		deposit_end_time = excluded.deposit_end_time,
+		voting_start_time = excluded.voting_start_time,
+		voting_end_time = excluded.voting_end_time`
 	_, err = db.SQL.Exec(proposalsQuery, proposalsParams...)
 	if err != nil {
 		return fmt.Errorf("error while storing proposals: %s", err)
@@ -202,9 +142,9 @@ INSERT INTO proposal(
 // GetProposal returns the proposal with the given id, or nil if not found
 func (db *Db) GetProposal(id uint64) (types.Proposal, error) {
 	var rows []*dbtypes.ProposalRow
-	err := db.Sqlx.Select(&rows, `SELECT * FROM proposal WHERE id = $1`, id)
+	err := db.SQL.Select(&rows, `SELECT * FROM proposal WHERE id = $1`, id)
 	if err != nil {
-		return types.Proposal{}, err
+		return types.Proposal{}, fmt.Errorf("error while getting proposal %d: %s", id, err)
 	}
 
 	if len(rows) == 0 {
@@ -213,28 +153,31 @@ func (db *Db) GetProposal(id uint64) (types.Proposal, error) {
 
 	row := rows[0]
 
-	var contentAny codectypes.Any
-	err = db.EncodingConfig.Codec.UnmarshalJSON([]byte(row.Content), &contentAny)
-	if err != nil {
-		return types.Proposal{}, err
-	}
+	trimContent := strings.TrimPrefix(row.Content, "{")
+	trimContent = strings.TrimPrefix(trimContent, "}")
+	jsonMessages := strings.Split(trimContent, ",")
 
-	var content govtypesv1beta1.Content
-	err = db.EncodingConfig.Codec.UnpackAny(&contentAny, &content)
-	if err != nil {
-		return types.Proposal{}, err
+	var messages []*codectypes.Any
+	for _, jsonMessage := range jsonMessages {
+		var msg codectypes.Any
+		err = db.Cdc.UnmarshalJSON([]byte(jsonMessage), &msg)
+		if err != nil {
+			return types.Proposal{}, err
+		}
+		messages = append(messages, &msg)
 	}
 
 	proposal := types.NewProposal(
 		row.ProposalID,
-		row.ProposalRoute,
-		row.ProposalType,
-		content,
+		row.Title,
+		row.Description,
+		row.Metadata,
+		messages,
 		row.Status,
 		row.SubmitTime,
 		row.DepositEndTime,
-		row.VotingStartTime,
-		row.VotingEndTime,
+		dbtypes.NullTimeToTime(row.VotingStartTime),
+		dbtypes.NullTimeToTime(row.VotingEndTime),
 		row.Proposer,
 	)
 	return proposal, nil
@@ -270,7 +213,7 @@ func (db *Db) UpdateProposal(update types.ProposalUpdate) error {
 		update.ProposalID,
 	)
 	if err != nil {
-		return fmt.Errorf("error while updating proposal: %s", err)
+		return fmt.Errorf("error while updating proposal %d: %s", update.ProposalID, err)
 	}
 
 	return nil
@@ -282,25 +225,29 @@ func (db *Db) SaveDeposits(deposits []types.Deposit) error {
 		return nil
 	}
 
-	query := `INSERT INTO proposal_deposit (proposal_id, depositor_address, amount, timestamp, height) VALUES `
+	query := `INSERT INTO proposal_deposit (proposal_id, depositor_address, 
+		amount, timestamp, transaction_hash, height) VALUES `
 	var param []interface{}
 	var accounts []types.Account
 	for i, deposit := range deposits {
-		vi := i * 5
-		query += fmt.Sprintf("($%d,$%d,$%d,$%d,$%d),", vi+1, vi+2, vi+3, vi+4, vi+5)
+		vi := i * 6
+
+		accounts = append(accounts, types.NewAccount(deposit.Depositor))
+
+		query += fmt.Sprintf("($%d,$%d,$%d,$%d,$%d,$%d),", vi+1, vi+2, vi+3, vi+4, vi+5, vi+6)
 		param = append(param, deposit.ProposalID,
 			deposit.Depositor,
 			pq.Array(dbtypes.NewDbCoins(deposit.Amount)),
 			deposit.Timestamp,
+			deposit.TransactionHash,
 			deposit.Height,
 		)
-		accounts = append(accounts, types.NewAccount(deposit.Depositor))
 	}
 
-	// Store the depositor account
+	// Store depositors accounts
 	err := db.SaveAccounts(accounts)
 	if err != nil {
-		return fmt.Errorf("error while storing depositor account: %s", err)
+		return fmt.Errorf("error while storing depositors accounts: %s", err)
 	}
 
 	query = query[:len(query)-1] // Remove trailing ","
@@ -323,10 +270,10 @@ WHERE proposal_deposit.height <= excluded.height`
 // SaveVote allows to save for the given height and the message vote
 func (db *Db) SaveVote(vote types.Vote) error {
 	query := `
-INSERT INTO proposal_vote (proposal_id, voter_address, option, timestamp, height)
-VALUES ($1, $2, $3, $4, $5)
+INSERT INTO proposal_vote (proposal_id, voter_address, option, weight, timestamp, height)
+VALUES ($1, $2, $3, $4, $5, $6)
 ON CONFLICT ON CONSTRAINT unique_vote DO UPDATE
-	SET option = excluded.option,
+	SET weight = excluded.weight, 
 		timestamp = excluded.timestamp,
 		height = excluded.height
 WHERE proposal_vote.height <= excluded.height`
@@ -337,9 +284,9 @@ WHERE proposal_vote.height <= excluded.height`
 		return fmt.Errorf("error while storing voter account: %s", err)
 	}
 
-	_, err = db.SQL.Exec(query, vote.ProposalID, vote.Voter, vote.Option.String(), vote.Timestamp, vote.Height)
+	_, err = db.SQL.Exec(query, vote.ProposalID, vote.Voter, vote.Option.String(), vote.Weight, vote.Timestamp, vote.Height)
 	if err != nil {
-		return fmt.Errorf("error while storing vote: %s", err)
+		return fmt.Errorf("error while storing vote for proposal %d: %s", vote.ProposalID, err)
 	}
 
 	return nil
@@ -390,17 +337,19 @@ func (db *Db) SaveProposalStakingPoolSnapshot(snapshot types.ProposalStakingPool
 	stmt := `
 INSERT INTO proposal_staking_pool_snapshot (proposal_id, bonded_tokens, not_bonded_tokens, height)
 VALUES ($1, $2, $3, $4)
-ON CONFLICT ON CONSTRAINT unique_staking_pool_snapshot DO UPDATE SET
-	proposal_id = excluded.proposal_id,
-    bonded_tokens = excluded.bonded_tokens,
-	not_bonded_tokens = excluded.not_bonded_tokens,
+ON CONFLICT ON CONSTRAINT unique_staking_pool_snapshot DO UPDATE 
+	SET bonded_tokens = excluded.bonded_tokens,
+	not_bonded_tokens = excluded.not_bonded_tokens, 
 	height = excluded.height
 WHERE proposal_staking_pool_snapshot.height <= excluded.height`
 
 	_, err := db.SQL.Exec(stmt,
-		snapshot.ProposalID, snapshot.Pool.BondedTokens.String(), snapshot.Pool.NotBondedTokens.String(), snapshot.Pool.Height)
+		snapshot.ProposalID, snapshot.Pool.BondedTokens.String(),
+		snapshot.Pool.NotBondedTokens.String(), snapshot.Pool.Height)
+
 	if err != nil {
-		return fmt.Errorf("error while storing proposal staking pool snapshot: %s", err)
+		return fmt.Errorf("error while storing proposal staking pool snapshot for proposal %d: %s",
+			snapshot.ProposalID, err)
 	}
 
 	return nil
@@ -413,7 +362,7 @@ func (db *Db) SaveProposalValidatorsStatusesSnapshots(snapshots []types.Proposal
 	}
 
 	stmt := `
-INSERT INTO proposal_validator_status_snapshot(proposal_id, validator_address, voting_power, status, jailed, height)
+INSERT INTO proposal_validator_status_snapshot (proposal_id, validator_address, voting_power, status, jailed, height) 
 VALUES `
 
 	var args []interface{}
@@ -428,17 +377,15 @@ VALUES `
 
 	stmt = stmt[:len(stmt)-1]
 	stmt += `
-ON CONFLICT ON CONSTRAINT unique_validator_status_snapshot DO UPDATE
-	SET proposal_id = excluded.proposal_id,
-		validator_address = excluded.validator_address,
-		voting_power = excluded.voting_power,
-		status = excluded.status,
+ON CONFLICT ON CONSTRAINT unique_validator_status_snapshot DO UPDATE 
+	SET voting_power = excluded.voting_power, 
+		status = excluded.status, 
 		jailed = excluded.jailed,
 		height = excluded.height
 WHERE proposal_validator_status_snapshot.height <= excluded.height`
 	_, err := db.SQL.Exec(stmt, args...)
 	if err != nil {
-		return fmt.Errorf("error while storing proposal validator statuses snapshot: %s", err)
+		return fmt.Errorf("error while storing proposal validator statuses snapshots: %s", err)
 	}
 
 	return nil
@@ -460,7 +407,7 @@ WHERE software_upgrade_plan.height <= excluded.height`
 	_, err := db.SQL.Exec(stmt,
 		proposalID, plan.Name, plan.Height, plan.Info, height)
 	if err != nil {
-		return fmt.Errorf("error while storing software upgrade plan: %s", err)
+		return fmt.Errorf("error while storing software upgrade plan for proposal %d: %s", proposalID, err)
 	}
 
 	return nil
@@ -472,7 +419,7 @@ func (db *Db) DeleteSoftwareUpgradePlan(proposalID uint64) error {
 
 	_, err := db.SQL.Exec(stmt, proposalID)
 	if err != nil {
-		return fmt.Errorf("error while deleting software upgrade plan: %s", err)
+		return fmt.Errorf("error while deleting software upgrade plan for proposal %d: %s", proposalID, err)
 	}
 
 	return nil