+++ modules/gov/module.go
@@ -7,7 +7,7 @@ import (
 
 	govsource "github.com/forbole/bdjuno/v4/modules/gov/source"
 
-	"github.com/forbole/juno/v4/modules"
+	"github.com/forbole/juno/v5/modules"
 )
 
 var (
@@ -19,21 +19,18 @@ var (
 
 // Module represent x/gov module
 type Module struct {
-	cdc             codec.Codec
-	db              *database.Db
-	source          govsource.Source
-	authModule      AuthModule
-	distrModule     DistrModule
-	inflationModule InflationModule
-	mintModule      MintModule
-	slashingModule  SlashingModule
-	stakingModule   StakingModule
+	cdc            codec.Codec
+	db             *database.Db
+	source         govsource.Source
+	distrModule    DistrModule
+	mintModule     MintModule
+	slashingModule SlashingModule
+	stakingModule  StakingModule
 }
 
 // NewModule returns a new Module instance
 func NewModule(
 	source govsource.Source,
-	authModule AuthModule,
 	distrModule DistrModule,
 	inflationModule InflationModule,
 	mintModule MintModule,
@@ -43,15 +40,13 @@ func NewModule(
 	db *database.Db,
 ) *Module {
 	return &Module{
-		cdc:             cdc,
-		source:          source,
-		authModule:      authModule,
-		distrModule:     distrModule,
-		inflationModule: inflationModule,
-		mintModule:      mintModule,
-		slashingModule:  slashingModule,
-		stakingModule:   stakingModule,
-		db:              db,
+		cdc:            cdc,
+		source:         source,
+		distrModule:    distrModule,
+		mintModule:     mintModule,
+		slashingModule: slashingModule,
+		stakingModule:  stakingModule,
+		db:             db,
 	}
 }