+++ modules/types/sources.go
@@ -2,36 +2,43 @@ package types
 
 import (
 	"fmt"
+	"os"
 
-	inflationtypes "github.com/evmos/evmos/v13/x/inflation/types"
-
-	"github.com/cosmos/cosmos-sdk/simapp/params"
-	"github.com/forbole/juno/v4/node/remote"
+	"cosmossdk.io/simapp/params"
+	"github.com/cometbft/cometbft/libs/log"
+	stakingkeeper "github.com/cosmos/cosmos-sdk/x/staking/keeper"
+	evmosapp "github.com/evmos/evmos/v15/app"
+	inflationtypes "github.com/evmos/evmos/v15/x/inflation/types"
+	"github.com/forbole/juno/v5/node/remote"
 
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
 	govtypesv1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
-	govtypesv1beta1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1beta1"
 	minttypes "github.com/cosmos/cosmos-sdk/x/mint/types"
 	slashingtypes "github.com/cosmos/cosmos-sdk/x/slashing/types"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
-	"github.com/forbole/juno/v4/node/local"
+	"github.com/forbole/juno/v5/node/local"
 
-	nodeconfig "github.com/forbole/juno/v4/node/config"
+	nodeconfig "github.com/forbole/juno/v5/node/config"
 
 	banksource "github.com/forbole/bdjuno/v4/modules/bank/source"
+	localbanksource "github.com/forbole/bdjuno/v4/modules/bank/source/local"
 	remotebanksource "github.com/forbole/bdjuno/v4/modules/bank/source/remote"
 	distrsource "github.com/forbole/bdjuno/v4/modules/distribution/source"
 	remotedistrsource "github.com/forbole/bdjuno/v4/modules/distribution/source/remote"
 	govsource "github.com/forbole/bdjuno/v4/modules/gov/source"
+	localgovsource "github.com/forbole/bdjuno/v4/modules/gov/source/local"
 	remotegovsource "github.com/forbole/bdjuno/v4/modules/gov/source/remote"
 	inflationsource "github.com/forbole/bdjuno/v4/modules/inflation/source"
+	localinflationsource "github.com/forbole/bdjuno/v4/modules/inflation/source/local"
 	remoteinflationsource "github.com/forbole/bdjuno/v4/modules/inflation/source/remote"
 	mintsource "github.com/forbole/bdjuno/v4/modules/mint/source"
 	remotemintsource "github.com/forbole/bdjuno/v4/modules/mint/source/remote"
 	slashingsource "github.com/forbole/bdjuno/v4/modules/slashing/source"
+	localslashingsource "github.com/forbole/bdjuno/v4/modules/slashing/source/local"
 	remoteslashingsource "github.com/forbole/bdjuno/v4/modules/slashing/source/remote"
 	stakingsource "github.com/forbole/bdjuno/v4/modules/staking/source"
+	localstakingsource "github.com/forbole/bdjuno/v4/modules/staking/source/local"
 	remotestakingsource "github.com/forbole/bdjuno/v4/modules/staking/source/remote"
 )
 
@@ -50,62 +57,56 @@ func BuildSources(nodeCfg nodeconfig.Config, encodingConfig *params.EncodingConf
 	case *remote.Details:
 		return buildRemoteSources(cfg)
 	case *local.Details:
-		return nil, fmt.Errorf("local mode is currently not supported")
+		return buildLocalSources(cfg, encodingConfig)
 
 	default:
 		return nil, fmt.Errorf("invalid configuration type: %T", cfg)
 	}
 }
 
-// func buildLocalSources(cfg *local.Details, encodingConfig *params.EncodingConfig) (*Sources, error) {
-// 	source, err := local.NewSource(cfg.Home, encodingConfig)
-// 	if err != nil {
-// 		return nil, err
-// 	}
+func buildLocalSources(cfg *local.Details, encodingConfig *params.EncodingConfig) (*Sources, error) {
+	source, err := local.NewSource(cfg.Home, encodingConfig)
+	if err != nil {
+		return nil, err
+	}
 
-// 	app := simapp.NewSimApp(
-// 		log.NewTMLogger(log.NewSyncWriter(os.Stdout)), source.StoreDB, nil, true, map[int64]bool{},
-// 		cfg.Home, 0, simapp.MakeTestEncodingConfig(), simapp.EmptyAppOptions{},
-// 	)
+	app := evmosapp.NewEvmos(
+		log.NewTMLogger(log.NewSyncWriter(os.Stdout)), source.StoreDB, nil, true, map[int64]bool{},
+		cfg.Home, 0, params.EncodingConfig{}, nil, nil,
+	)
+	sources := &Sources{
+		BankSource: localbanksource.NewSource(source, banktypes.QueryServer(app.BankKeeper)),
+		// DistrSource:     localdistrsource.NewSource(source, distrtypes.QueryServer(app.DistrKeeper)),
+		GovSource:       localgovsource.NewSource(source, govtypesv1.QueryServer(app.GovKeeper)),
+		InflationSource: localinflationsource.NewSource(source, inflationtypes.QueryServer(app.InflationKeeper)),
+		// MintSource:      localmintsource.NewSource(source, minttypes.QueryServer(app.MintKeeper)),
+		SlashingSource: localslashingsource.NewSource(source, slashingtypes.QueryServer(app.SlashingKeeper)),
+		StakingSource:  localstakingsource.NewSource(source, stakingkeeper.Querier{Keeper: &app.StakingKeeper}),
+	}
 
-// 	evmosApp := evmosapp.NewEvmos(
-// 		log.NewTMLogger(log.NewSyncWriter(os.Stdout)), source.StoreDB, nil, true, map[int64]bool{},
-// 		cfg.Home, 0, simapp.MakeTestEncodingConfig(), simapp.EmptyAppOptions{},
-// 	)
+	// Mount and initialize the stores
+	err = source.MountKVStores(app, "keys")
+	if err != nil {
+		return nil, err
+	}
 
-// 	sources := &Sources{
-// 		BankSource:      localbanksource.NewSource(source, banktypes.QueryServer(app.BankKeeper)),
-// 		DistrSource:     localdistrsource.NewSource(source, distrtypes.QueryServer(app.DistrKeeper)),
-// 		GovSource:       localgovsource.NewSource(source, govtypesv1.QueryServer(evmosApp.GovKeeper), nil),
-// 		InflationSource: localinflationsource.NewSource(source, inflationtypes.QueryServer(evmosApp.InflationKeeper)),
-// 		MintSource:      localmintsource.NewSource(source, minttypes.QueryServer(app.MintKeeper)),
-// 		SlashingSource:  localslashingsource.NewSource(source, slashingtypes.QueryServer(app.SlashingKeeper)),
-// 		StakingSource:   localstakingsource.NewSource(source, stakingkeeper.Querier{Keeper: app.StakingKeeper}),
-// 	}
+	err = source.MountTransientStores(app, "tkeys")
+	if err != nil {
+		return nil, err
+	}
 
-// 	// Mount and initialize the stores
-// 	err = source.MountKVStores(app, "keys")
-// 	if err != nil {
-// 		return nil, err
-// 	}
+	err = source.MountMemoryStores(app, "memKeys")
+	if err != nil {
+		return nil, err
+	}
 
-// 	err = source.MountTransientStores(app, "tkeys")
-// 	if err != nil {
-// 		return nil, err
-// 	}
+	err = source.InitStores()
+	if err != nil {
+		return nil, err
+	}
 
-// 	err = source.MountMemoryStores(app, "memKeys")
-// 	if err != nil {
-// 		return nil, err
-// 	}
-
-// 	err = source.InitStores()
-// 	if err != nil {
-// 		return nil, err
-// 	}
-
-// 	return sources, nil
-// }
+	return sources, nil
+}
 
 func buildRemoteSources(cfg *remote.Details) (*Sources, error) {
 	source, err := remote.NewSource(cfg.GRPC)
@@ -116,7 +117,7 @@ func buildRemoteSources(cfg *remote.Details) (*Sources, error) {
 	return &Sources{
 		BankSource:      remotebanksource.NewSource(source, banktypes.NewQueryClient(source.GrpcConn)),
 		DistrSource:     remotedistrsource.NewSource(source, distrtypes.NewQueryClient(source.GrpcConn)),
-		GovSource:       remotegovsource.NewSource(source, govtypesv1.NewQueryClient(source.GrpcConn), govtypesv1beta1.NewQueryClient(source.GrpcConn)),
+		GovSource:       remotegovsource.NewSource(source, govtypesv1.NewQueryClient(source.GrpcConn)),
 		InflationSource: remoteinflationsource.NewSource(source, inflationtypes.NewQueryClient(source.GrpcConn)),
 		MintSource:      remotemintsource.NewSource(source, minttypes.NewQueryClient(source.GrpcConn)),
 		SlashingSource:  remoteslashingsource.NewSource(source, slashingtypes.NewQueryClient(source.GrpcConn)),