+++ types/gov.go
@@ -3,121 +3,27 @@ package types
 import (
 	"time"
 
+	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+
 	govtypesv1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
-	govtypesv1beta1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1beta1"
+	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 )
 
 const (
 	ProposalStatusInvalid = "PROPOSAL_STATUS_INVALID"
 )
 
-// DepositParams contains the data of the deposit parameters of the x/gov module
-type DepositParams struct {
-	MinDeposit       sdk.Coins `json:"min_deposit,omitempty" yaml:"min_deposit"`
-	MaxDepositPeriod int64     `json:"max_deposit_period,omitempty" yaml:"max_deposit_period"`
-}
-
-// NewDepositParam allows to build a new DepositParams
-func NewDepositParam(d *govtypesv1.DepositParams) DepositParams {
-	return DepositParams{
-		MinDeposit:       d.MinDeposit,
-		MaxDepositPeriod: d.MaxDepositPeriod.Nanoseconds(),
-	}
-}
-
-// NewGenesisDepositParam allows to build a new DepositParams
-func NewGenesisDepositParam(d *govtypesv1beta1.DepositParams) DepositParams {
-	return DepositParams{
-		MinDeposit:       d.MinDeposit,
-		MaxDepositPeriod: d.MaxDepositPeriod.Nanoseconds(),
-	}
-}
-
-// VotingParams contains the voting parameters of the x/gov module
-type VotingParams struct {
-	VotingPeriod int64 `json:"voting_period,omitempty" yaml:"voting_period"`
-}
-
-// NewVotingParams allows to build a new VotingParams instance
-func NewVotingParams(v *govtypesv1.VotingParams) VotingParams {
-	return VotingParams{
-		VotingPeriod: v.VotingPeriod.Nanoseconds(),
-	}
-}
-
-// NewGenesisVotingParams allows to build a new VotingParams instance
-func NewGenesisVotingParams(v *govtypesv1beta1.VotingParams) VotingParams {
-	return VotingParams{
-		VotingPeriod: v.VotingPeriod.Nanoseconds(),
-	}
-}
-
 // GovParams contains the data of the x/gov module parameters
 type GovParams struct {
-	DepositParams DepositParams `json:"deposit_params" yaml:"deposit_params"`
-	VotingParams  VotingParams  `json:"voting_params" yaml:"voting_params"`
-	TallyParams   TallyParams   `json:"tally_params" yaml:"tally_params"`
-	Height        int64         `json:"height" ymal:"height"`
+	*govtypesv1.Params
+	Height int64 `json:"height" ymal:"height"`
 }
 
-// GenesisGovParams contains the data of the x/gov module parameters
-type GenesisGovParams struct {
-	DepositParams DepositParams      `json:"deposit_params" yaml:"deposit_params"`
-	VotingParams  VotingParams       `json:"voting_params" yaml:"voting_params"`
-	TallyParams   GenesisTallyParams `json:"tally_params" yaml:"tally_params"`
-	Height        int64              `json:"height" ymal:"height"`
-}
-
-// TallyParams contains the tally parameters of the x/gov module
-type TallyParams struct {
-	Quorum        string `json:"quorum,omitempty"`
-	Threshold     string `json:"threshold,omitempty"`
-	VetoThreshold string `json:"veto_threshold,omitempty" yaml:"veto_threshold"`
-}
-
-// GenesisTallyParams contains genesis tally parameters of the x/gov module
-type GenesisTallyParams struct {
-	Quorum        sdk.Dec `json:"quorum,omitempty"`
-	Threshold     sdk.Dec `json:"threshold,omitempty"`
-	VetoThreshold sdk.Dec `json:"veto_threshold,omitempty" yaml:"veto_threshold"`
-}
-
-// NewTallyParams allows to build a new TallyParams instance
-func NewTallyParams(t *govtypesv1.TallyParams) TallyParams {
-	return TallyParams{
-		Quorum:        t.Quorum,
-		Threshold:     t.Threshold,
-		VetoThreshold: t.VetoThreshold,
-	}
-}
-
-// NewGenesisTallyParams allows to build a new GenesisTallyParams instance
-func NewGenesisTallyParams(t *govtypesv1beta1.TallyParams) GenesisTallyParams {
-	return GenesisTallyParams{
-		Quorum:        t.Quorum,
-		Threshold:     t.Threshold,
-		VetoThreshold: t.VetoThreshold,
-	}
-}
-
-// NewGovParams allows to build a new GovParams instance
-func NewGovParams(votingParams VotingParams, depositParams DepositParams, tallyParams TallyParams, height int64) *GovParams {
+func NewGovParams(params *govtypesv1.Params, height int64) *GovParams {
 	return &GovParams{
-		DepositParams: depositParams,
-		VotingParams:  votingParams,
-		TallyParams:   tallyParams,
-		Height:        height,
-	}
-}
-
-// NewGenesisGovParams allows to build a new GenesisGovParams instance
-func NewGenesisGovParams(votingParams VotingParams, depositParams DepositParams, tallyParams GenesisTallyParams, height int64) *GenesisGovParams {
-	return &GenesisGovParams{
-		DepositParams: depositParams,
-		VotingParams:  votingParams,
-		TallyParams:   tallyParams,
-		Height:        height,
+		Params: params,
+		Height: height,
 	}
 }
 
@@ -125,36 +31,39 @@ func NewGenesisGovParams(votingParams VotingParams, depositParams DepositParams,
 
 // Proposal represents a single governance proposal
 type Proposal struct {
-	ProposalRoute   string
-	ProposalType    string
-	ProposalID      uint64
-	Content         govtypesv1beta1.Content
+	ID              uint64
+	Title           string
+	Summary         string
+	Metadata        string
+	Messages        []*codectypes.Any
 	Status          string
 	SubmitTime      time.Time
 	DepositEndTime  time.Time
-	VotingStartTime time.Time
-	VotingEndTime   time.Time
+	VotingStartTime *time.Time
+	VotingEndTime   *time.Time
 	Proposer        string
 }
 
 // NewProposal return a new Proposal instance
 func NewProposal(
 	proposalID uint64,
-	proposalRoute string,
-	proposalType string,
-	content govtypesv1beta1.Content,
+	title string,
+	summary string,
+	metadata string,
+	messages []*codectypes.Any,
 	status string,
 	submitTime time.Time,
 	depositEndTime time.Time,
-	votingStartTime time.Time,
-	votingEndTime time.Time,
+	votingStartTime *time.Time,
+	votingEndTime *time.Time,
 	proposer string,
 ) Proposal {
 	return Proposal{
-		Content:         content,
-		ProposalRoute:   proposalRoute,
-		ProposalType:    proposalType,
-		ProposalID:      proposalID,
+		ID:              proposalID,
+		Title:           title,
+		Summary:         summary,
+		Metadata:        metadata,
+		Messages:        messages,
 		Status:          status,
 		SubmitTime:      submitTime,
 		DepositEndTime:  depositEndTime,
@@ -164,32 +73,16 @@ func NewProposal(
 	}
 }
 
-// Equal tells whether p and other contain the same data
-func (p Proposal) Equal(other Proposal) bool {
-	return p.ProposalRoute == other.ProposalRoute &&
-		p.ProposalType == other.ProposalType &&
-		p.ProposalID == other.ProposalID &&
-		p.Content.String() == other.Content.String() &&
-		p.Status == other.Status &&
-		p.SubmitTime.Equal(other.SubmitTime) &&
-		p.DepositEndTime.Equal(other.DepositEndTime) &&
-		p.VotingStartTime.Equal(other.VotingStartTime) &&
-		p.VotingEndTime.Equal(other.VotingEndTime) &&
-		p.Proposer == other.Proposer
-}
-
 // ProposalUpdate contains the data that should be used when updating a governance proposal
 type ProposalUpdate struct {
 	ProposalID      uint64
 	Status          string
-	VotingStartTime time.Time
-	VotingEndTime   time.Time
+	VotingStartTime *time.Time
+	VotingEndTime   *time.Time
 }
 
 // NewProposalUpdate allows to build a new ProposalUpdate instance
-func NewProposalUpdate(
-	proposalID uint64, status string, votingStartTime, votingEndTime time.Time,
-) ProposalUpdate {
+func NewProposalUpdate(proposalID uint64, status string, votingStartTime, votingEndTime *time.Time) ProposalUpdate {
 	return ProposalUpdate{
 		ProposalID:      proposalID,
 		Status:          status,
@@ -202,11 +95,12 @@ func NewProposalUpdate(
 
 // Deposit contains the data of a single deposit made towards a proposal
 type Deposit struct {
-	ProposalID uint64
-	Depositor  string
-	Amount     sdk.Coins
-	Timestamp  time.Time
-	Height     int64
+	ProposalID      uint64
+	Depositor       string
+	Amount          sdk.Coins
+	Timestamp       time.Time
+	TransactionHash string
+	Height          int64
 }
 
 // NewDeposit return a new Deposit instance
@@ -215,14 +109,16 @@ func NewDeposit(
 	depositor string,
 	amount sdk.Coins,
 	timestamp time.Time,
+	transactionHash string,
 	height int64,
 ) Deposit {
 	return Deposit{
-		ProposalID: proposalID,
-		Depositor:  depositor,
-		Amount:     amount,
-		Timestamp:  timestamp,
-		Height:     height,
+		ProposalID:      proposalID,
+		Depositor:       depositor,
+		Amount:          amount,
+		Timestamp:       timestamp,
+		TransactionHash: transactionHash,
+		Height:          height,
 	}
 }
 
@@ -233,6 +129,7 @@ type Vote struct {
 	ProposalID uint64
 	Voter      string
 	Option     govtypesv1.VoteOption
+	Weight     string
 	Timestamp  time.Time
 	Height     int64
 }
@@ -242,6 +139,7 @@ func NewVote(
 	proposalID uint64,
 	voter string,
 	option govtypesv1.VoteOption,
+	weight string,
 	timestamp time.Time,
 	height int64,
 ) Vote {
@@ -249,6 +147,7 @@ func NewVote(
 		ProposalID: proposalID,
 		Voter:      voter,
 		Option:     option,
+		Weight:     weight,
 		Timestamp:  timestamp,
 		Height:     height,
 	}
@@ -309,7 +208,7 @@ type ProposalValidatorStatusSnapshot struct {
 	ProposalID           uint64
 	ValidatorConsAddress string
 	ValidatorVotingPower int64
-	ValidatorStatus      int
+	ValidatorStatus      stakingtypes.BondStatus
 	ValidatorJailed      bool
 	Height               int64
 }
@@ -319,7 +218,7 @@ func NewProposalValidatorStatusSnapshot(
 	proposalID uint64,
 	validatorConsAddr string,
 	validatorVotingPower int64,
-	validatorStatus int,
+	validatorStatus stakingtypes.BondStatus,
 	validatorJailed bool,
 	height int64,
 ) ProposalValidatorStatusSnapshot {